backend/requirements.txt
# Core Backend Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2

# Production Server (for WebSocket support)
gunicorn==23.0.0
eventlet==0.40.3

# Optional: Development Dependencies (uncomment if needed)
# pytest==7.4.3
# pytest-flask==1.3.0
# black==23.11.0
# flake8==6.1.0
